# Conflicts:
#	src/minebombers/Game.cpp

+++ src/minebombers/Map.cpp
@@ -68,7 +68,7 @@ void Map::setTileset(Tileset& tileset) {
 }
 
 bool Map::floorAt(sf::Vector2u newPosition) {
-    if (newPosition.x >= 0 && newPosition.x < getSize().x && newPosition.y >= 0 && newPosition.y < getSize().y) {
+    if ((int) newPosition.x >= 0 && newPosition.x < getSize().x && (int) newPosition.y >= 0 && newPosition.y < getSize().y) {
         TileType tileType = this->getTile(newPosition).getType();
         return newPosition.x > 0 && newPosition.y > 0
                && newPosition.x <= width && newPosition.y <= height
